# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "proviant is a simple and intuitive application to track your bought products and their expiration date to prevent waste of food.",
        "title": "proviant",
        "contact": {
            "name": "Isotop7",
            "url": "https://codeberg.org/isotop7/proviant",
            "email": "hendrik@hr94.de"
        },
        "license": {
            "name": "MIT",
            "url": "https://codeberg.org/isotop7/proviant/-/blob/main/LICENSE"
        },
        "version": "0.2.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/api/v1/products/{id}/open": {
            "post": {
                "description": "Sets the OpenedAt timestamp on a product. Returns 409 with the existing OpenedAt if the product is already opened, prompting the frontend to confirm. Re-submit with `force: true` to overwrite.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Mark product as opened",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Force overwrite flag",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/v1.OpenProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/v1.OpenProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "database.Product": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "barcode": {
                    "type": "string"
                },
                "categories": {
                    "type": "string"
                },
                "co2KgPerKg": {
                    "type": "number"
                },
                "countries": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "daysAfterOpening": {
                    "type": "integer"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "expireAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isPrivate": {
                    "type": "boolean"
                },
                "minStockAmount": {
                    "type": "integer"
                },
                "notificationLeadDays": {
                    "type": "integer"
                },
                "notifiedAt": {
                    "type": "string"
                },
                "openedAt": {
                    "type": "string"
                },
                "priceOverride": {
                    "type": "number"
                },
                "productName": {
                    "type": "string"
                },
                "removalReason": {
                    "type": "string"
                },
                "scannedAt": {
                    "type": "string"
                },
                "storageLocation": {
                    "$ref": "#/definitions/database.StorageLocation"
                },
                "storageLocationId": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "database.StorageLocation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "householdId": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "v1.OpenProductRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "type": "boolean"
                }
            }
        },
        "v1.OpenProductResponse": {
            "type": "object",
            "properties": {
                "openedAt": {
                    "type": "string"
                },
                "product": {},
                "triggeredBy": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JSON Web Token": {
            "type": "basic"
        }
    },
...
}